Extracted basic dom into submodule which keeps the nativescript and svelte specifics apart for clarity. We now have more control over the nodes which should let us create a FrameElement for navigation
