+++ src/dom/basicdom/Readme.txt
@@ -1,4 +1,4 @@
-This dom implementation was originally from the nativescript-vue project and is governed by the LICENSE file in this directory.
+This dom implementation was extracted from and based on the nativescript-vue project and is governed by the LICENSE file in this directory.
 https://github.com/nativescript-vue/nativescript-vue/tree/master/platform/nativescript/renderer
 or more accurately
 https://github.com/nativescript-vue/nativescript-vue/tree/6d7b16915e2ba1548398065edeae76a74cabbee2/platform/nativescript/renderer
