+++ src/dom/basicdom/element-registry.ts
@@ -1,5 +1,4 @@
-import { View } from 'tns-core-modules/ui/core/view'
-import { normalizeElementName, ComponentMeta } from './ViewNode'
+import { normalizeElementName } from './ViewNode'
 import ElementNode from './ElementNode';
 
 export type NodeResolver = () => ElementNode;
@@ -16,10 +15,6 @@ function registerElementResolver(elementName: string, entry: ElementEntry) {
     throw new Error(`Element for ${normalizedName} already registered.`)
   }
   elementMap[normalizedName] = entry
-}
-
-export function registerElement(elementName:string, resolver: ()=>typeof View, meta: ComponentMeta = null) {
-  registerElementResolver(elementName, { resolver: () => new ElementNode(elementName, resolver(), meta) });
 }
 
 export function registerCustomElementNode(elementName:string, resolver: NodeResolver) {
