+++ src/dom/basicdom/index.ts
@@ -1,0 +1,7 @@
+export { default as DocumentNode } from "./DocumentNode"
+export { default as ElementNode } from "./ElementNode"
+export { default as CommentNode } from "./CommentNode"
+export { default as PropertyNode } from "./PropertyNode"
+export { default as ViewNode, normalizeElementName } from "./ViewNode"
+export { default as TextNode } from "./TextNode"
+export { registerCustomElementNode, createElement } from "./element-registry"
