+++ src/dom/style-scope.d.ts
