+++ src/index.ts
@@ -1,7 +1,6 @@
 import { Frame } from 'tns-core-modules/ui/frame'
 import { run, on, launchEvent } from 'tns-core-modules/application'
-import { DocumentNode, ViewNode, createElement } from './dom'
-import { registerNativeElements } from './nativescript-elements'
+import { ViewNode, NativeElementNode, createElement, initializeDom } from './dom';
 
 declare global {
     export class SvelteComponent {
@@ -10,64 +9,23 @@ declare global {
     }
 }
 
-function installGlobalShims() {
-   
-   //expose our fake dom as global document for svelte components
-   let window = global as any;
-
-   window.window = global;
-   window.document = new DocumentNode();
-
-   window.requestAnimationFrame = (action: ()=>{}) => {
-       setTimeout(action, 33); //about 30 fps
-   }
-
-   window.getComputedStyle = (node: ViewNode) => {
-       return node.nativeView.style;
-   }
-
-   window.performance = {
-       now() {
-           return Date.now();
-       }
-   };
-
-   window.CustomEvent = class {
-       detail: any;
-       eventName: string;
-       type: string;
-       constructor(name: string, detail: any = null) {
-          this.eventName = name; //event name for nativescript
-          this.type = name; // type for svelte
-          this.detail = detail;
-       }
-   }
-
-   
-
-}
-
-
 export function svelteNative(startPage: typeof SvelteComponent, data: any) {
-    registerNativeElements();
-    installGlobalShims();
-
-    const document = (global as any).document as DocumentNode;
+    const document = initializeDom();
 
     //our application main navigation frame
-    let frame = createElement('frame');
+    let frame = createElement('frame') as NativeElementNode;
     document.appendChild(frame);
-    
+
     //wait for launch
-    on(launchEvent, ()=>{
+    on(launchEvent, () => {
         let page = new startPage({
             target: frame,
             props: data || {}
         });
         //dirty way to find page's native view
-        (frame.nativeView as Frame).navigate({ create: () => frame.firstElement().nativeView});
+        (frame.nativeView as Frame).navigate({ create: () => (frame.firstElement() as NativeElementNode).nativeView });
     })
 
-    run({create: () => frame.nativeView});
+    run({ create: () => frame.nativeView });
 }
