+++ src/svelte-components.d.ts
@@ -2,3 +2,4 @@
 declare module "*.html" {
     export default SvelteComponent;
 }
+
